Moved more batch-protocols, improved naming and docstrings
