+++ tianshou/policy/imitation/gail.py
@@ -5,10 +5,9 @@ import torch
 import torch.nn.functional as F
 
 from tianshou.data import ReplayBuffer, to_numpy, to_torch
-from tianshou.data.types import RolloutBatchProtocol
+from tianshou.data.types import LogpOldProtocol, RolloutBatchProtocol
 from tianshou.policy import PPOPolicy
 from tianshou.policy.modelfree.pg import TDistParams
-from tianshou.policy.modelfree.ppo import BatchWithLogpOldProtocol
 
 
 class GAILPolicy(PPOPolicy):
@@ -105,7 +104,7 @@ class GAILPolicy(PPOPolicy):
 
     def process_fn(
         self, batch: RolloutBatchProtocol, buffer: ReplayBuffer, indices: np.ndarray
-    ) -> BatchWithLogpOldProtocol:
+    ) -> LogpOldProtocol:
         """Pre-process the data from the provided replay buffer.
 
         Used in :meth:`update`. Check out :ref:`process_fn` for more information.
